+++ plugin_ide.ui/src/com/googlecode/goclipse/ui/navigator/GoNavigatorContentProvider.java
@@ -116,8 +116,7 @@ public class GoNavigatorContentProvider extends AbstractNavigatorContentProvider
 	
 	@Override
 	protected void addBuildTargetsContainer(IProject project, ArrayList2<Object> projectChildren) {
-		// Temporarily disabled
-		//super.addBuildTargetsContainer(project, projectChildren);
+		super.addBuildTargetsContainer(project, projectChildren);
 	}
 	
 	@Override