+++ plugin_tooling/src-util/melnorme/utilbox/core/CommonException.java
@@ -2,6 +2,9 @@ package melnorme.utilbox.core;
 
 import static melnorme.utilbox.core.Assert.AssertNamespace.assertNotNull;
 
+import melnorme.lang.tooling.data.StatusException;
+import melnorme.lang.tooling.data.StatusLevel;
+
 /**
  * A generic status exception.
  * Has an associated message, and optionally an associated Exception cause.
@@ -16,6 +19,14 @@ public class CommonException extends Exception {
 	
 	public CommonException(String message, Throwable cause) {
 		super(assertNotNull(message), cause);
+	}
+	
+	public StatusException toStatusException(StatusLevel statusLevel) {
+		if(this instanceof StatusException) {
+			return (StatusException) this;
+		} else {
+			return new StatusException(statusLevel, getMessage());
+		}
 	}
 	
 }